Refactor move the code that normalize input parameters into DataProcessor class
